Error handling
Была реализована валидация запросов, при помощи FluentValidation для
классов добавления, удаления, редактирования и получения. Также был
реализован класс, обрабатывающий ошибки, как в developer mode, так и в
production.

+++ Application/Subjects/Delete.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
-using Domain;
+using Application.Core;
 using MediatR;
 using Persistence;
 
@@ -9,24 +9,30 @@ namespace Application.Subjects
 {
   public class Delete
   {
-    public class Command : IRequest
+    public class Command : IRequest<Result<Unit>>
     {
       public Guid Id { get; set; }
     }
-    public class Handler : IRequestHandler<Command>
+    public class Handler : IRequestHandler<Command, Result<Unit>>
     {
       private readonly DataContext _context;
 
       public Handler(DataContext context) =>
         _context = context;
 
-      public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
+      public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
       {
         var subject = await _context.Subjects.FindAsync(request.Id);
 
+        if (subject == null)
+          return (null);
         _context.Subjects.Remove(subject);
-        await _context.SaveChangesAsync();
-        return (Unit.Value);
+
+        var result = await _context.SaveChangesAsync() > 0;
+
+        if (!result)
+          return (Result<Unit>.Failure("Ошибка в удалении занятия"));
+        return (Result<Unit>.Success(Unit.Value));
       }
     }
   }
